requirements.txt
fastapi~=0.111.0
uvicorn~=0.30.1
pydantic~=2.7.3
pymongo~=4.7.3
pydantic-mongo~=2.3.0
google-cloud-storage~=2.16.0
python-multipart~=0.0.9